+++ ZXingObjC/client/result/ZXProductResultParser.m
@@ -27,7 +27,7 @@
     return nil;
   }
   NSString *rawText = [ZXResultParser massagedText:result];
-  int length = [rawText length];
+  NSUInteger length = [rawText length];
   for (int x = 0; x < length; x++) {
     unichar c = [rawText characterAtIndex:x];
     if (c < '0' || c > '9') {
