+++ ZXingObjC/datamatrix/ZXDataMatrixWriter.m
@@ -67,7 +67,7 @@
   //1. step: Data encodation
   NSString *encoded = [ZXHighLevelEncoder encodeHighLevel:contents shape:shape minSize:minSize maxSize:maxSize];
 
-  ZXSymbolInfo *symbolInfo = [ZXSymbolInfo lookup:encoded.length shape:shape minSize:minSize maxSize:maxSize fail:YES];
+  ZXSymbolInfo *symbolInfo = [ZXSymbolInfo lookup:(int)encoded.length shape:shape minSize:minSize maxSize:maxSize fail:YES];
 
   //2. step: ECC generation
   NSString *codewords = [ZXDataMatrixErrorCorrection encodeECC200:encoded symbolInfo:symbolInfo];
