+++ ZXingObjC/multi/qrcode/detector/ZXMultiFinderPatternFinder.m
@@ -52,7 +52,7 @@ float const DIFF_MODSIZE_CUTOFF = 0.5f;
  */
 - (NSArray *)selectBestPatternsWithError:(NSError **)error {
   NSMutableArray *_possibleCenters = [NSMutableArray arrayWithArray:[self possibleCenters]];
-  int size = [_possibleCenters count];
+  NSUInteger size = [_possibleCenters count];
 
   if (size < 3) {
     if (error) *error = NotFoundErrorInstance();
